Add element position detection
Replace the existing implementation with a re-written component with support for position detection.

The new component supports detecting either element size, position, or both - depending on which callback props are provided.

The new default behavior matches the previous component with the exception that the `onResize` callback is a full `DOMRect` instance, not an object with only `width` and `height`. This change has the potential to cause issues for consumers who are allowing the returned object to be serialized to JSON as a `DOMRect` instance will always return and empty object when serialized. Returning the raw `DOMRect` instance is a conscious choice to keep performance as fast as possible.