Support react 16
`findDOMNode` behaves differently in React 16 due to the switch to fiber under the hood and returns `null` when it shouldn't.

I can't find a proper reference to what has changed. The official documentation hasn't been updated to reflect any breaking changes. It's best to simply avoid using `findDOMNode` in this scenario.